anti-human-league/solutions/solutions/requirements.txt
requests
beautifulsoup4
pytesseract
Pillow
requests-toolbelt
paramiko
